i've created a class just to conform to the protocols and now the object is no longer the main point of the gravity, is more generic way, now i can put anykind of object
