+++ PhysicsWorld/PhysicsWorld/App/Assets.xcassets/AppIcon.appiconset/Contents.json
@@ -1,17 +1,13 @@
 {
   "images" : [
     {
-      "idiom" : "universal",
-      "platform" : "ios",
-      "size" : "1024x1024"
-    },
-    {
       "appearances" : [
         {
           "appearance" : "luminosity",
           "value" : "dark"
         }
       ],
+      "filename" : "aaaa.png",
       "idiom" : "universal",
       "platform" : "ios",
       "size" : "1024x1024"
