--- PhysicsWorld/PhysicsWorld/Presenters/Enviroment/CustomTriangle.swift
@@ -1,39 +1,0 @@
-//
-//  Triangle.swift
-//  Budget
-//
-//  Created by Jairo Júnior on 26/02/24.
-//
-
-//
-
-import SwiftUI
-
-//struct CustomTriangle: ObjectsProtocol, Shape{
-//    var position: CGPoint
-//    var velocity: CGPoint
-//    var mass: Double
-//    var isDynamic: Bool
-//    
-//    init(position: CGPoint = .zero, velocity: CGPoint = .zero, mass: Double = .zero, isStatic: Bool = true) {
-//        self.position = position
-//        self.velocity = velocity
-//        self.mass = mass
-//        self.isDynamic = isStatic
-//    }
-//    
-//    func path(in rect: CGRect) -> Path {
-//        var path = Path()
-//        
-//        path.move(to: .init(x: rect.midX, y: rect.minY))
-//        path.addLine(to: .init(x: rect.maxX, y: rect.maxY))
-//        path.addLine(to: .init(x: rect.minX, y: rect.maxY))
-//        
-//        
-//        return path
-//    }
-//}
-
-//#Preview {
-//    CustomTriangle()
-//}
